Added related projects section in README
Hi, I was inspired by your project and I created IntelliSpace, which brings the same concept of VSpaceCode to the IntelliJ platform.
I think that VSpaceCode users may be interested also in IntelliSpace and so I added a "related projects" section on the bottom of the README in which there is a link to my project. Of course you can add other projects like ours if you know some.
I hope that this is not too invasive.