Remove buggy `when` keyword in commands
The `when` keyword has been introduced in VSCode v1.27 [1], but it is supported
by _VSCode_ keybindings **only**, and it's not supported by _VSCodeVim_ as a
quick search on their repository shows [2].

Because both commands were called every time, this was triggering an error
notification on both OS when `SPC f f` was used. I know this default will break
the behaviour on MacOS, but I'm not aware of any solution to support both
platform for the moment.

The problem is already explained in the README with the workarounds, so nothing
more to do here.

[1] https://code.visualstudio.com/updates/v1_27#_platform-specific-keybindings
[2] https://github.com/VSCodeVim/Vim/search?q=isMac&type=